Artemisinin resistance based on molecular marker in the Kelch 13 gene

The discovery of molecular markers of resistance in the Kelch 13 gene gives us the unusual opportunity to monitor the emergence and spread of resistance nearly in real time. Assessing the prevalence of these molecular markers in other regions can provide an early warning system to trigger rapid responses to develop control strategies.

artemisinin

Format

A tibble with 2434 rows and 16 columns:

Variable	Description

continent	Name of continent

val	Effect on clearance

country	Name of country

year	Year

tested	Number of cases tested

prev	Prevalence of Kelch 13 gene molecular marker

lon	Longitude coordinate for site location

title	Title of paper from which data was derived

pubMedId	PubMed identifier

sid	Site identifier

mutation	Type of mutation

site	Name of site

estLoc	Location

present	Number of cases where mutation was present

lat	Latitude coordinate for site location

authors	Name of authors of paper from which data was derived

Source

https://www.wwarn.org/tracking-resistance/artemisinin-molecular-surveyor

Details

The Artemisinin Molecular Surveyor is an interactive map that summarises the prevalence of these molecular markers in the propeller region of the Kelch 13 gene of the malaria parasite by location and year, providing a clear and standardised visualisation of the evidence to date to inform key international, regional and national monitoring strategies.

Examples

artemisinin

#> # A tibble: 2,434 x 16
#>    continent val   country year  tested prev  lon   title pubMedId sid  
#>    <chr>     <chr> <chr>   <chr> <chr>  <chr> <chr> <chr> <chr>    <chr>
#>  1 Asia      clea… Myanmar 2014  62     96.7… 92.8… Spre… 25704894 PSHHG
#>  2 Asia      clea… Myanmar 2013  30     63.3… 97.3… Spre… 25704894 PSHHG
#>  3 Asia      clea… Myanmar 2014  40     62.5… 97.3… Spre… 25704894 PSHHG
#>  4 Asia      clea… Myanmar 2013  188    42.0… 98.1… Spre… 25704894 PSHHG
#>  5 Asia      clea… Myanmar 2014  99     72.7… 98.1… Spre… 25704894 PSHHG
#>  6 Asia      clea… Myanmar 2013  3      66.6… 97.8… Spre… 25704894 PSHHG
#>  7 Asia      clea… Myanmar 2014  5      60.0… 97.8… Spre… 25704894 PSHHG
#>  8 Asia      clea… Myanmar 2013  19     89.4… 94.0… Spre… 25704894 PSHHG
#>  9 Asia      clea… Myanmar 2014  10     100.… 94.0… Spre… 25704894 PSHHG
#> 10 Asia      clea… Bangla… 2014  25     100.… 92.1… Spre… 25704894 PSHHG
#> # … with 2,424 more rows, and 6 more variables: mutation <chr>, site <chr>,
#> #   estLoc <chr>, present <chr>, lat <chr>, authors <chr>