Antimalarial medicine quality

Substandard and falsified (SF) antimalarial medicines pose an immediate danger to vulnerable populations and could increase the threat of malaria drug resistance emerging and spreading. The major challenge in preventing the spread and distribution of poor quality medical products is the lack of accessible and reliable information on their prevalence.

medicine_quality

Format

A tibble with 1553 rows and 37 columns:

Source

https://www.wwarn.org/about-us/medicine-quality/mq-scientific-literature-surveyor-antimalarial-quality-data

Examples

medicine_quality

#> # A tibble: 1,553 x 37
#>    type  publication_id doi   title first_author  year publisher
#>    <chr>          <dbl> <chr> <chr> <chr>        <dbl> <chr>    
#>  1 Anti…              8 10.1… Subs… Tipke, M.     2008 Malaria …
#>  2 Anti…              8 10.1… Subs… Tipke, M.     2008 Malaria …
#>  3 Anti…              8 10.1… Subs… Tipke, M.     2008 Malaria …
#>  4 Anti…              8 10.1… Subs… Tipke, M.     2008 Malaria …
#>  5 Anti…              8 10.1… Subs… Tipke, M.     2008 Malaria …
#>  6 Anti…              8 10.1… Subs… Tipke, M.     2008 Malaria …
#>  7 Anti…              8 10.1… Subs… Tipke, M.     2008 Malaria …
#>  8 Anti…              8 10.1… Subs… Tipke, M.     2008 Malaria …
#>  9 Anti…              8 10.1… Subs… Tipke, M.     2008 Malaria …
#> 10 Anti…              2 http… Impa… Keoluangkho…  2008 The Amer…
#> # … with 1,543 more rows, and 30 more variables: publication_type <dbl>,
#> #   publication_type_name <chr>, url <chr>, inn_drug_id <dbl>,
#> #   short_name <chr>, drug_name <chr>, countryId <dbl>, countryName <chr>,
#> #   lat <dbl>, lon <dbl>, id <dbl>, location_name <chr>, latitude <dbl>,
#> #   longitude <dbl>, survey_data_id <dbl>, outlet_type_id <dbl>,
#> #   outlet_type_name <chr>, sampling_type_id <dbl>, sampling_type_name <chr>,
#> #   drug_quality_category_id <dbl>, drug_quality_category_name <chr>,
#> #   failure_rate <dbl>, starting_date <chr>, ending_date <chr>,
#> #   samples_total <dbl>, name <chr>, sdq_id <dbl>, sdq_num_sample <dbl>,
#> #   sdq_category <chr>, samples_total_unit <chr>