Plasmodium vivax clinical trials

Data on prevalence of Plasmodium vivax antimalarial clinical trials across the world. The data are derived from a recent review of Plasmodium vivax antimalarial clinical trials.

vivax

Format

A tibble with 16326 rows and 44 columns:

Source

https://www.wwarn.org/tracking-resistance/vivax-surveyor

Examples

vivax

#> # A tibble: 16,326 x 44
#>    article_id title first_author journal year_published abstract_content doi  
#>         <int> <chr> <chr>        <chr>   <chr>          <chr>            <chr>
#>  1       3945 Indi… Akpede GO    Annals… 1992           A total of 522 … ""   
#>  2       3140 Comp… Abhyankar AV Viral … 2006           Dengue virus in… ""   
#>  3        525 An o… Gould, L Ha… Transa… 2008           From September … ""   
#>  4       1282 Salm… Lee, W S     Journa… 1999           OBJECTIVE: To r… ""   
#>  5       1241 Meli… Chaowagul, … The Jo… 1989           In a prospectiv… ""   
#>  6        654 Clin… Cheesbrough… The Pe… 1997           BACKGROUND: Inv… ""   
#>  7       1321 Eosi… Chan, YC     Journa… 2004           A 49-year-old w… ""   
#>  8       4037 Coxi… Botros BA    The Jo… 1995           Retrospective s… ""   
#>  9       2774 Orga… Ameen S      Asian … 2012           Background: The… ""   
#> 10       4222 Tren… Singhal L    Indian… 2014           Enteric fever i… ""   
#> # … with 16,316 more rows, and 37 more variables: sample_source <chr>,
#> #   study_type_name <chr>, category_name <chr>, pubMedId <int>, url <chr>,
#> #   study_start_year <int>, study_end_year <int>, estimated_year <chr>,
#> #   age_group_label <chr>, estimated_age <chr>, max_age <dbl>,
#> #   max_age_months <chr>, min_age <dbl>, min_age_months <chr>, family <chr>,
#> #   genus <chr>, group_name <chr>, kingdom <chr>, microorganism_mapping <chr>,
#> #   species <chr>, organism_name <chr>, organism.snomed_concept_name <chr>,
#> #   clinical_disorder_name <chr>, clinical_disorder.snomed_concept_name <chr>,
#> #   positive <int>, tested <int>, percentage <chr>, study_id <int>,
#> #   country_name <chr>, iso_3166_1_alpha_3 <chr>, google_lat <dbl>,
#> #   google_lon <dbl>, province_name <chr>, site_name <chr>, country_id <int>,
#> #   region <chr>, sub_region <chr>